test: separate pure unit tests from Testbench integration tests
