+++ tests/Integration/WorkbenchIntegrationTest.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Neo4j\LaravelBoost\Tests\Unit;
+namespace Neo4j\LaravelBoost\Tests\Integration;
 
-use Neo4j\LaravelBoost\Tests\TestCase;
+use Neo4j\LaravelBoost\Tests\WorkbenchTestCase;
 use Orchestra\Testbench\Concerns\WithWorkbench;
 
 /**
  * Ensures tests use {@see WithWorkbench} + root testbench.yaml
  * (web route discovery), not only plain PHPUnit.
  */
-class WorkbenchIntegrationTest extends TestCase
+class WorkbenchIntegrationTest extends WorkbenchTestCase
 {
     public function test_workbench_web_route_from_testbench_yaml_is_reachable(): void
     {
